support ffmpeg-v6.0

+++ android/ijkplayer/ijkplayer-java/src/main/java/tv/danmaku/ijk/media/player/IjkEventListener.java
@@ -34,5 +34,5 @@ public interface IjkEventListener {
 
 
 
-    void onEvent(IjkMediaPlayer mp, int what, int arg1, int arg2, Object extra);
+    void onEvent(IjkMediaPlayer mp, int what, long arg1, long arg2, Object extra);
 }
